+++ frontend/src/components/BudgetLineItems/SummaryCardsSection/SummaryCardsSection.js
@@ -1,6 +1,6 @@
 import RoundedBox from "../../UI/RoundedBox";
 
-const BudgetLineTotalSummaryCard = (title) => {
+const BudgetLineTotalSummaryCard = ({ title }) => {
     return (
         <RoundedBox className="padding-y-205 padding-x-4 padding-right-9 ">
             <h3 className="margin-0 margin-bottom-3 font-12px text-base-dark text-normal">{title}</h3>
@@ -11,8 +11,8 @@ const BudgetLineTotalSummaryCard = (title) => {
 const SummaryCardsSection = () => {
     return (
         <div className="display-flex flex-justify">
+            <BudgetLineTotalSummaryCard title="TODO: Replace me" />
             <BudgetLineTotalSummaryCard title="Budget Lines Total" />
-            <BudgetLineTotalSummaryCard />
         </div>
     );
 };
