+++ frontend/src/components/UI/README.md
@@ -1,0 +1,5 @@
+# UI
+
+This directory contains the DOM primitive components that all other components will be built on.
+
+For example, we may build a `AllBudgetLinesTable` by composing from from the `Table`component found here.
